Refactored Category and Product models: Extracted relationships and scopes into separate traits
