+++ app/Models/Product.php
@@ -6,10 +6,11 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\SoftDeletes;
 use App\Enums\ProductStatus;
+use App\Models\Traits\ProductRelationshipsTrait;
 
 class Product extends Model
 {
-    use HasFactory, SoftDeletes;
+    use HasFactory, SoftDeletes, ProductRelationshipsTrait;
 
     protected $fillable = [
         'product_name',
@@ -45,19 +46,4 @@ class Product extends Model
     protected $casts = [
         'status' => ProductStatus::class,
     ];
-
-    public function category()
-    {
-        return $this->belongsTo(Category::class, 'category_id', 'id');
-    }
-
-    public function brand()
-    {
-        return $this->belongsTo(Brand::class, 'brand_id', 'id');
-    }
-
-    public function subcategory()
-    {
-        return $this->belongsTo(Category::class, 'subcategory_id', 'id');
-    }
 }
