+++ app/Models/Traits/CategoryRelationshipsTrait.php
@@ -1,0 +1,26 @@
+<?php
+
+namespace App\Models\Traits;
+
+use Illuminate\Database\Eloquent\Relations\HasMany;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use App\Models\Category;
+use App\Models\Product;
+
+trait CategoryRelationshipsTrait
+{
+    public function parentCategory(): BelongsTo
+    {
+        return $this->belongsTo(Category::class, 'parent_id');
+    }
+
+    public function subcategories(): HasMany
+    {
+        return $this->hasMany(Category::class, 'parent_id');
+    }
+
+    public function products(): HasMany
+    {
+        return $this->hasMany(Product::class);
+    }
+}
