+++ app/Models/Traits/CategoryScopesTrait.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace App\Models\Traits;
+
+use Illuminate\Database\Eloquent\Builder;
+
+trait CategoryScopesTrait
+{
+    public function scopeByParent(Builder $query, $parentId): Builder
+    {
+        return $query->where('parent_id', $parentId);
+    }
+}
