+++ app/Models/Traits/ProductRelationshipsTrait.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace App\Models\Traits;
+
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use App\Models\Category;
+use App\Models\Brand;
+
+trait ProductRelationshipsTrait
+{
+    public function category()
+    {
+        return $this->belongsTo(Category::class, 'category_id', 'id');
+    }
+
+    public function brand()
+    {
+        return $this->belongsTo(Brand::class, 'brand_id', 'id');
+    }
+
+    public function subcategory()
+    {
+        return $this->belongsTo(Category::class, 'subcategory_id', 'id');
+    }
+}
